Working Turrets & Multiplayer Backend Update
- Large code clean up
- Updated to NGO Preview 7, preview 8 has laggy client network transforms.
- Falcon now has working turrets that fires at ships

+++ Assets/Scripts/Multiplayer/Player/PlayerManagerMP.cs
@@ -4,6 +4,7 @@ using UnityEngine;
 using Unity.Netcode;
 using System.Text;
 using System;
+using Unity.Mathematics;
 
 namespace MultiplayerRunTime
 {
@@ -130,6 +131,7 @@ namespace MultiplayerRunTime
         {
             if (ShipSpawned)
             {
+                
                 LocalSpaceship.transform.position = position;
                 LocalSpaceship.transform.forward = forward;
                 FindObjectOfType<MouseFlightControllerMP>().frozenDirection = forward;
@@ -142,7 +144,7 @@ namespace MultiplayerRunTime
             UserMenu menu = PasswordLobbyMP.Singleton.menu;
             menu.ShowInfoOverlay(true);
             menu.infoPopUp.UpperLabel = "Starting in...";
-            menu.infoPopUp.LowerLabel = time.ToString("F0");
+            menu.infoPopUp.LowerLabel = ((float)time).ToString("F0");
             if (time <= 3)
             {
                 menu.infoPopUp.MakeRed();
